Re-land 5929 (r5882) with crash fix and also this change:
http://codereview.chromium.org/11392/show

which fixes the remainder of the UI test issues apparently.

http://crbug.com/4620


git-svn-id: svn://svn.chromium.org/chrome/trunk/src@5954 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/tab_contents.h
@@ -393,8 +393,8 @@ class TabContents : public PageNavigator,
   void RemoveInfoBar(InfoBarDelegate* delegate);
   
   // Enumeration and access functions.
-  size_t infobar_delegate_count() const { return infobar_delegates_.size(); }
-  InfoBarDelegate* GetInfoBarDelegateAt(size_t index) {
+  int infobar_delegate_count() const { return infobar_delegates_.size(); }
+  InfoBarDelegate* GetInfoBarDelegateAt(int index) {
     return infobar_delegates_.at(index);
   }
